+++ packages/provider-utils/src/index.ts
@@ -15,19 +15,18 @@ export * from './load-api-key';
 export { loadOptionalSetting } from './load-optional-setting';
 export { loadSetting } from './load-setting';
 export * from './parse-json';
+export { parseJsonEventStream } from './parse-json-event-stream';
 export { parseProviderOptions } from './parse-provider-options';
 export * from './post-to-api';
 export * from './remove-undefined-entries';
 export * from './resolve';
 export * from './response-handler';
+export { asSchema, jsonSchema } from './schema';
+export type { Schema } from './schema';
+export type { ToolCall } from './types/tool-call';
+export type { ToolResult, ToolResultContent } from './types/tool-result';
 export * from './uint8-utils';
 export * from './validate-types';
 export * from './validator';
 export * from './without-trailing-slash';
-
-export { asSchema, jsonSchema } from './schema';
-export type { Schema } from './schema';
 export { zodSchema } from './zod-schema';
-
-export type { ToolCall } from './types/tool-call';
-export type { ToolResult, ToolResultContent } from './types/tool-result';
